+++ src/Simulation/Simulators/QuantumSimulator/ApplyControlledZ.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using Microsoft.Quantum.Simulation.Core;
 using Microsoft.Quantum.Intrinsic.Interfaces;
 
@@ -12,7 +13,7 @@ namespace Microsoft.Quantum.Simulation.Simulators
         {
             this.CheckQubits(new QArray<Qubit>(new Qubit[]{ control, target }));
 
-            MCZ(this.Id, 1, new uint[]{(uint)control.Id}, (uint)target.Id);
+            MCZ(this.Id, 1, new IntPtr[]{(IntPtr)control.Id}, (IntPtr)target.Id);
         }
     }
 }
