+++ src/Simulation/Simulators/QuantumSimulator/ApplyUncontrolledSWAP.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using Microsoft.Quantum.Simulation.Core;
 using Microsoft.Quantum.Intrinsic.Interfaces;
 
@@ -16,9 +17,9 @@ namespace Microsoft.Quantum.Simulation.Simulators
 
             this.CheckQubits(new QArray<Qubit>(new Qubit[]{ qubit1, qubit2 }));
 
-            MCX(this.Id, 1, new uint[]{(uint)qubit1.Id}, (uint)qubit2.Id);
-            MCX(this.Id, 1, new uint[]{(uint)qubit2.Id}, (uint)qubit1.Id);
-            MCX(this.Id, 1, new uint[]{(uint)qubit1.Id}, (uint)qubit2.Id);
+            MCX(this.Id, 1, new IntPtr[]{(IntPtr)qubit1.Id}, (IntPtr)qubit2.Id);
+            MCX(this.Id, 1, new IntPtr[]{(IntPtr)qubit2.Id}, (IntPtr)qubit1.Id);
+            MCX(this.Id, 1, new IntPtr[]{(IntPtr)qubit1.Id}, (IntPtr)qubit2.Id);
         }
     }
 }
