+++ src/Simulation/Simulators/QuantumSimulator/M.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using Microsoft.Quantum.Simulation.Core;
 using Microsoft.Quantum.Intrinsic.Interfaces;
 
@@ -13,7 +14,7 @@ namespace Microsoft.Quantum.Simulation.Simulators
             this.CheckQubit(target);
             //setting qubit as measured to allow for release
             target.IsMeasured = true;
-            return M(this.Id, (uint)target.Id).ToResult();
+            return M(this.Id, (IntPtr)target.Id).ToResult();
         }
     }
 }
