+++ src/Simulation/Simulators/QuantumSimulator/NativeImports.cs
@@ -10,87 +10,87 @@ namespace Microsoft.Quantum.Simulation.Simulators
     public partial class QuantumSimulator
     {
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "allocateQubit")]
-        private static extern void AllocateOne(uint id, uint qubit_id);
+        private static extern void AllocateOne(uint id, IntPtr qubit_id);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "release")]
-        private static extern bool ReleaseOne(uint id, uint qubit_id);
+        private static extern bool ReleaseOne(uint id, IntPtr qubit_id);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "Exp")]
-        private static extern void Exp(uint id, uint n, Pauli[] paulis, double angle, uint[] ids);
+        private static extern void Exp(uint id, uint n, Pauli[] paulis, double angle, IntPtr[] ids);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCExp")]
-        private static extern void MCExp(uint id, uint n, Pauli[] paulis, double angle, uint nc, uint[] ctrls, uint[] ids);
+        private static extern void MCExp(uint id, uint n, Pauli[] paulis, double angle, uint nc, IntPtr[] ctrls, IntPtr[] ids);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "H")]
-        private static extern void H(uint id, uint qubit);
+        private static extern void H(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCH")]
-        private static extern void MCH(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCH(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "M")]
-        private static extern uint M(uint id, uint q);
+        private static extern uint M(uint id, IntPtr q);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "Measure")]
-        private static extern uint Measure(uint id, uint n, Pauli[] b, uint[] ids);
+        private static extern uint Measure(uint id, uint n, Pauli[] b, IntPtr[] ids);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "R")]
-        private static extern void R(uint id, Pauli basis, double angle, uint qubit);
+        private static extern void R(uint id, Pauli basis, double angle, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCR")]
-        private static extern void MCR(uint id, Pauli basis, double angle, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCR(uint id, Pauli basis, double angle, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "S")]
-        private static extern void S(uint id, uint qubit);
+        private static extern void S(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "AdjS")]
-        private static extern void AdjS(uint id, uint qubit);
+        private static extern void AdjS(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCS")]
-        private static extern void MCS(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCS(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCAdjS")]
-        private static extern void MCAdjS(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCAdjS(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "T")]
-        private static extern void T(uint id, uint qubit);
+        private static extern void T(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "AdjT")]
-        private static extern void AdjT(uint id, uint qubit);
+        private static extern void AdjT(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCT")]
-        private static extern void MCT(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCT(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCAdjT")]
-        private static extern void MCAdjT(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCAdjT(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "X")]
-        private static extern void X(uint id, uint qubit);
+        private static extern void X(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCX")]
-        private static extern void MCX(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCX(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "Y")]
-        private static extern void Y(uint id, uint qubit);
+        private static extern void Y(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCY")]
-        private static extern void MCY(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCY(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "Z")]
-        private static extern void Z(uint id, uint qubit);
+        private static extern void Z(uint id, IntPtr qubit);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "MCZ")]
-        private static extern void MCZ(uint id, uint count, uint[] ctrls, uint qubit);
+        private static extern void MCZ(uint id, uint count, IntPtr[] ctrls, IntPtr qubit);
 
-        private delegate bool DumpCallback(uint idx, double real, double img);
+        private delegate bool DumpCallback(IntPtr idx, double real, double img);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "Dump")]
         private static extern void sim_Dump(uint id, DumpCallback callback);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "DumpQubits")]
-        private static extern bool sim_DumpQubits(uint id, uint cout, uint[] ids, DumpCallback callback);
+        private static extern bool sim_DumpQubits(uint id, uint cout, IntPtr[] ids, DumpCallback callback);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "JointEnsembleProbability")]
-        private static extern double JointEnsembleProbability(uint id, uint n, Pauli[] b, uint[] q);
+        private static extern double JointEnsembleProbability(uint id, uint n, Pauli[] b, IntPtr[] q);
 
         [DllImport(QSIM_DLL_NAME, ExactSpelling = true, CallingConvention = CallingConvention.Cdecl, EntryPoint = "random_choice")]
         private static extern Int64 random_choice(uint id, Int64 size, double[] p);
